Add decode input class and its capi wrapper.
